requirements.txt
django
django-rest-framework
django-cors-headers
httpx
grpcio
grpcio-tools